Squashed commit of the following:
commit 5aabf62c0fa55aab2affc7285aedf9b25ba0d83f
Author: Luke Benstead <kazade@gmail.com>
Date:   Sun Dec 29 20:13:37 2019 +0000

    Revert "Enable debug logging in tests"

    This reverts commit 127078d30af5db751e8ae15ea4c8619a7cc49bbb.

commit 4f1ef5f854c0fb8ce13804af6accc4fe0756a0cb
Author: Luke Benstead <kazade@gmail.com>
Date:   Sun Dec 29 20:13:23 2019 +0000

    Revert some things

commit cb5072cc88a317c32a19ae58c87b66e1b524ae86
Author: Luke Benstead <kazade@gmail.com>
Date:   Sun Dec 29 19:39:49 2019 +0000

    Handle exceptions during window creation by terminating

commit 89190fced1c77d526bef98f7f5889d285e322787
Author: Luke Benstead <kazade@gmail.com>
Date:   Sun Dec 29 18:33:40 2019 +0000

    Try again

commit 6e0ac47cc9c5d2c45352f33655a6c7d59d231ad6
Author: Luke Benstead <kazade@gmail.com>
Date:   Sun Dec 29 16:39:25 2019 +0000

    Try again

commit ce766bef419aed0121c5ace7b6f2afbbbde2dbbe
Author: Luke Benstead <kazade@gmail.com>
Date:   Sun Dec 29 15:56:30 2019 +0000

    Try again

commit 43d1bf528fc87a6848f07edaa796e5c7447f8096
Author: Luke Benstead <kazade@gmail.com>
Date:   Sun Dec 29 15:01:59 2019 +0000

    Make sure we install lxdream first to get the deps

commit 4a11e411723c932661c794cdcd78ef5f9d03c773
Author: Luke Benstead <kazade@gmail.com>
Date:   Sun Dec 29 14:58:56 2019 +0000

    Try different lxdream version

commit 90a298885297d746ac06adc376a64ebb7b6fd9c8
Author: Luke Benstead <kazade@gmail.com>
Date:   Sun Dec 29 14:39:29 2019 +0000

    Try manual install of a later lxdream build

commit a127afe6f980aefba0eca11bcdcdbbcb6df7e699
Merge: 900a9213 cfc785f2
Author: Luke Benstead <kazade@gmail.com>
Date:   Sun Dec 29 14:04:02 2019 +0000

    Merge branch 'master' into update-ci

commit 900a921391f278da855cdf1043c2569318f4d9c1
Author: Luke Benstead <kazade@gmail.com>
Date:   Sun Dec 29 14:03:58 2019 +0000

    Add some logging

commit defc207b053e8f69c591efa499841105da398e37
Author: Luke Benstead <kazade@gmail.com>
Date:   Sun Dec 29 13:33:07 2019 +0000

    Try again with -b

commit 921461a40fcd04dff4c0210ec2fdb449bfd468f7
Author: Luke Benstead <kazade@gmail.com>
Date:   Sun Dec 29 12:30:08 2019 +0000

    Try update

commit 9cdb63d658a891f1c441c6632cc81db78adc6e40
Author: Luke Benstead <kazade@gmail.com>
Date:   Sun Dec 29 10:13:54 2019 +0000

    Remove -b flag

commit 160f872f50e9bbd4f2e29907c83e9568e4d43e11
Author: Luke Benstead <kazade@gmail.com>
Date:   Sun Dec 29 09:58:20 2019 +0000

    Try using the .cdi to run Dreamcast tests

+++ simulant/debug.h
@@ -82,7 +82,7 @@ private:
         DebugElementType type = DET_LINE;
         Colour colour = Colour::WHITE;
         bool depth_test = true;
-        double duration = 0.0;
+        float duration = 0.0;
 
         smlt::Vec3 points[2]; // For lines, or the first one for points
         float size; // Diameter for spheres + points
